requirements.txt
streamlit
groq
python-dotenv
gTTS
SpeechRecognition
pillow
audio-recorder-streamlit
